Replaced most instances where the code was checking to see if an object was of type StringType to a check to see if it is String Type or UnicodeType.
I left the instances where it looked like the StringType check should be
separate (for instance, when converting a string to unicode).

The new version of CherryPy ensures that all parameters are unicode and
thus a lot of these functions were failing because they did not check
for it and StringType is not a parent class.