Removed references to LmServer objects LMObject and LMError
LmBackend should not reference things in LmServer as compute nodes do not have LmServer code.  We might want to consider moving a common exception class to LmBackend as we have one in LmServer and LmCompute.  I did not see any reason to make this object inherit from LMObj but we may want to reevaluate later.