Removed comments that were for session locking.
At this time, we are only issuing serial requests from the QGIS plugin and changing the locking
mechanism would really only benefit concurrent requests for the same logged in user.  CherryPy
seems to favor implicit locking at this time, so we will use it.  If the QGIS plugin changes
its access pattern in the future to issue more concurrent requests, we may want to address this
again.