Added filter for attributes
With the modification of the LmAttObj dir method, entries that used to just be attributes are added to the main member pool.  Then on serialization, the items are duplicated as tags and attributes in XML.  This causes problems in deserialization because multiple entries are seen and they are converted to a list with a plural tag.  Thus, the proper attribute cannot be referenced and there is duplication.  This filter removes those entries from the dir() determined list and serialization treats these attributes as XML attributes as it should.